requirements.txt
# This file is generated automatically by `make requirements.txt`.
# Do not edit manually.
python_version >= "3.10"
matplotlib==3.8.0
numpy==1.26.2
opencv_python==4.8.1.78
opencv_python_headless==4.8.1.78
Requests==2.31.0
scipy==1.11.4
trimesh==4.0.5
transforms3d==0.4.1
tqdm
pyyaml
paho-mqtt

#mesh-raycast
git+https://github.com/szabolcsdombi/python-mesh-raycast.git
